feat(#13): clean up gradle build, publish to artifactory
This commit cleans up the gradle build logic and moves most of the
tricky parts to buildSrc plugins, which hopefully can be re-used
across other projects.

We also modify the publishing config to publish to the artifactory
repo instead of s3.  This will allow us to share the artifacts
with early-access customers without requiring IAM auth, and in
a way that will be consistent for other SDK languages in the near
future.