feat: Add endpoint override for Momento
The plan of record is to make endpoint available in the authToken. SDK will be parsing the authToken to determine the endpoint and then prefix it with `control.` to perform control operations and `cache.` to perform cache operations.

Currently, the endpoint isn't available in the authToken. Hence an endpointOverride option is added. Momento will prefer the endpointOverride option over the endpoint in the authToken.

The change also switches Momento construction from init() to Builder pattern which is more consistent with other standard Java SDKs like those of other cloud providers.

Now the object construction will be:
```
Momento momento = Momento.builder()
                    .authToken("authToken")
                    .endpointOverride("developer-gautam-dev.preprod.a.momentohq.com")
                    .build();
```