feat: Make Cache Management simpler for Multi Cache customers
Currently the expectation from the customers is to have a Cache entity per each of their caches. Each cache entity is in turn tied to the GRPC channel to the SCS backend. So, we want customers to reuse their Cache entity. While this may be simpler for customers with a single cache, wherein they could simply create the Cache client once at the application start and inject into their code. This becomes significantly harder for customers with multiple caches since they now need to map cache names to correct cache clients and write their own registry. Simply creating a new cache client in application code will be very costly. Although the channel initialization etc. is handled lazily in the grpc layer usually first few cache operations can take longer.

There are two ways to solve this problem:

1. Just provide a CacheClient which now is more of client than an entity. This client can take in cache name on each SET/GET request and then pass those transactions to the backend. This does come with the risk of CacheNotFound and PermissionDenied as CacheClients aren't now Cache entity specific.

2. The second approach is to separate the entity from the actual GRPC layer. The reason why Cache entity is costly right now is because of its tight coupling with the Grpc layer. While the risk of CacheNotFound and PermissionDenied still remains, it is now cheap to request Cache entity on each call as the grpc client will be shared amongst all entities.

I have gone in with approach #2. But can easily be convinced to go to #1. I like #2 because for customers who prefer, inject these entities in their appropriate applications. But at the same time it is cheap enough for customers who don't prefer to do so.

```
   class CacheManager {
     private final Momento momento;

     @Inject
     public CacheManager(Momento momento) {
       this.momento = momento;
     }

     public void addToCache(String cacheName, CacheData cacheData) {
       Cache cache = momento.cacheBuilder(cacheName, cacheData.ttlSeconds).build();
       cache.set(cacheData.key, cacheData.value);
     }

     public Data getData(String cacheName, String cacheKey) {
       Cache cache = momento.cacheBuilder(cacheName, cacheData.ttlSeconds).build();
       retrun Data.create(cache.get(cacheKey));
     }
   }
```

A discussion around createIfDoesntExist -

The way it is implemented currently we would never want customers to use this operation in code unless they are going to reuse the Cache entity. Although it ensures that Cache will always be present, adding that option to production code will always result in a GET first before any other operation thus incurring additional latency.

We should revist this pattern if we remove this:

1. Customers will have to create caches before hand
2. Maintain a registry of Cache entities to make operations cheaper instead of getting Cache entities on the fly
3. Add custom code to catch NotFoundExceptions and then create cache

There are following other ways in which we can solve this problem:

1. Documentation
Ensure that we document the shortcomings of this method

2. Change the semantics
If customers provide this option, we do the cache existence check as a part of normal operation

e.g.
```
   public void set(String key, String value) {
     try {
       scsGrpcClient.set(key, value, ttl)
     } catch(CacheNotFoundException e) {
       if (createIfDoesntExist) {
         try {
           momento.createCache(cacheName);
         } catch (CacheAlreadyExistsException) {
           // ignore
         }
         set(key, value);
       }
     }
   }
```
There is still some race so some calls may end up taking longer around the time cache doesn't exist. But will eventually even out. We may also have to assess risks if any to control plane if multiple requests end up hitting it.