feat: Remove Optional from SetResponse values
Optional in  GetResponse makes sense that the values are Optional. Empty when there is a Miss and have values when there is a a Hit.

In context of SetResponse, this isn't true, if set succeeded there will always be a value in SetResponse if it didn't then there is no SetResponse.