chore: update getCache response (#155)
* chore: update getCache response

Update the response returned by the get and getAsync methods to match
the pattern we have established in other SDKs. The response has three
implementations: Hit, Miss, and Error, and it no longer throws. It is
intended to be used with pattern matching.

Change the deleteAsync method to delete instead of getting.

Add a test dependency on AssertJ, a fluent assertion library.

* Method names and misc improvements

* Fix tests

* change get response error to extend SdkException

Add a test utility method to generate a random string.

* remove extra getters from CacheGetResponse