chore: remove synchronous methods. (#157)
* chore: update getCache response

Update the response returned by the get and getAsync methods to match
the pattern we have established in other SDKs. The response has three
implementations: Hit, Miss, and Error, and it no longer throws. It is
intended to be used with pattern matching.

Change the deleteAsync method to delete instead of getting.

Add a test dependency on AssertJ, a fluent assertion library.

* Method names and misc improvements

* chore: remove synchronous methods.

Remove the synchronous versions of the cache methods. The async versions
can easily be made synchronous, and we reduce the code surface
significantly.

* remove 'async' from internal method names