chore: Remove member variable optionals and fix warnings
Remove member variable and method argument optionals. Java optionals are
sad compared to Kotlin's and provide no null guarantees, so it is valid
to supply a method that takes an optional with a null.

Add a jsr305 dependency to get access to @Nonnull and @Nullable
annotations.

Annotate the cache client constructor chain with @Nullable and @Nonnull.
These are the only places where the sdk can throw an exception if nulls
are provided.

Fix a few misc warnings.