PR fixes
Return a NavigableSet in the sorted set fetch response.

Delete the SortOrder enum because the return value of the sorted set
fetch methods can be viewed in either order.

Change ScoredElement from a generic class to one that can return a
String or a byte[] so that we only need one fetch elements signature.