feat: Add sortedSetGetRank (#235)
* feat: Add sortedSetGetRank

Add sortedSetGetRank. This returns the rank of the given sorted set
element.

Add generic code for executing a grpc function. This should cut down on
boilerplate in the internal data client.

* add order to sortedSetGetRank

Add order to sortedSetGetRank because it affects the ranks of the items
in the set. Defaults to ascending to match the node.js sdk. That feels
like a bad default to me because then a higher score has a higher rank.