chore: improve missing env var message, illustrate how to build examples against local SDK
This commit adds a bit more error handling to the EnvVarCredentialProvider so that if the
environment variable is missing, we will give a slightly more helpful error message.

Also includes a commented-out gradle snippet for the examples project which allows you to
compile the examples against your local copy of the SDK, so that it is easier to test changes
as you are developing.