feat: persistent store spike
Adds a POC Persistent Store client, complete with client interface,
client, driver program, and response types.

The design focus is around the get request and set response. Since the
persistent store allows users to store and retrieve data of various
types, we must handle those cases carefully.

To store data, we overload the `set` method. There is a `set` overload
for each of the supported value types.

To retrieve data, we have a single `GetResponse` class. This class
stores the value as an object and the type as reported from the server
as an enumerated field. Users can inspect the type of the value by
calling the accessor `getType` and then calling the corresponding
value accessor for strongly typed access (`getValueAsString`,
`getValueAsByteArray`, ...).

In the event a user calls the wrong method for the underlying type, we
throw an exception.

We add a driver program to demonstrate usage in
`PersistentStoreDriver`. This shows how to exhaustively check a get
response when the underlying type is unknown. We also show what to do
when the type is expected.