feat: persistent store spike (#358)
Adds a Preview Storage client.

To retrieve data, we have a single GetResponse class. This class
stores the value as a separate `Optional<StorageValue>` object.
If the get was not found, then this is empty. Otherwise present.

`StorageValue` allows users call `get[Type]` where `[Type]` is
one of the possible types. When the function call does not match
the underlying type, `Optional.empty()` is returned.