refactor: introduce `Store(Item)?NotFoundException` classes
This PR detects when a `NOT_FOUND` gRPC error was specific to the
storage client or not. For the storage client `NOT_FOUND`, there are
two cases: the store was not found and the item was not found.

To distinguish these two cases we use string matching on the
grpcException message.