fix: circular calling on `GetResponse::found`
Adds a new interface `GetResponseFound` to solve the previous
implementation's circular calling on `GetResponse::found`.

Previously
this method returned a `GetResponse.Found`, so on a response of this
type, a caller could write
`response.found().get().found().get().found()...` ad infinitum. This
was not desirable.

To solve this we introduce a new interface `GetResponseFound` which
`GetResponse.Found` implements. This way `GetResponse::found` returns
an `Optional.of` an object implementing that, which cuts the cycle.