feat: storage get response 2.0 (#375)
Split GetResponse.Success into GetResponse.Found and, in the event the item isn't found, GetResponse.NotFound. This sheds the optional on value.
Add a valueWhenFound shortcut accessor on the interface that returns a MomentoOptional<StorageValue>.
Throw exceptions when StorageValue::get* are called on the wrong type. This sheds the Optional on those responses. Return a MomentoOptional which allows for both optional behavior and custom error messages. A careful user can still switch on StorageValue::getType for exhaustiveness checks.
Add GetResponse::asFound which returns GetResponseFound in the event of a GetResponse.Found otherwise throws an exception. Will add this on demand per user feedback in order to keep the API slim.
Introduce MomentoOptional for the convenience of a parameter-less orElseThrow method. The orElseThrow method will throw a ClientSdkException when the optional is empty with an informative error message. This makes using GetResponse::found slightly less long winded.