feat: Prototype for concurrent operations limiting
Modify cache Get and Get to add optional concurrent operations limiting.
It works by creating an executor with a max number of threads equal to
the limit. Get and Set calls are given to the executor and wait on the
executor's internal queue until there is a thread free to take them,
implicitly limiting the number of concurrent requests.

Add a general ScsFutureStub method that takes a gRPC call, a gRPC to
Momento response converter, and an error handler. Something like this
should let us cut a lot of boilerplate out of the data client. It won't
work for the batch call, since that uses a different type of stub.