fix: fix failing storage tests (#390)
Stop looking at error metadata to determine whether an already exists
error is for a store or cache. We can check it when that is added on the
server side.

Update a test that was checking that the server could support empty keys
in the store client. That changed on the server side.