feat: Add a concurrency limiting executor to the ClientBase
Add an Executor to the ClientBase that will be used to limit the number
of concurrent calls that a Momento client can make.

Add closing logic for the executor service.

Add an abstract doClose method that replaces the close method previously
used by the child classes, since the base class now implements close.