feat: add the read concern header
Add the read concern option to the cache Configuration and thread it
through to the gRPC internals of the client.

Add an argument to the UserHeaderInterceptor constructor representing a
list of extra headers to add to every request. The stubs manager uses
this to add the read concern header if it isn't set to balanced.

Make the base cache test class create clients with both balanced and
consistent read concern, and set the cache client used by all the tests
to the consistent one if the CONSISTENT_READS env var is defined.

Add tasks to run integration tests for the different clients.

Add make targets to use the new tasks.