feat: add the read concern header (#408)
* feat: add the read concern header

Add the read concern option to the cache Configuration and thread it
through to the gRPC internals of the client.

Add an argument to the UserHeaderInterceptor constructor representing a
list of extra headers to add to every request. The stubs manager uses
this to add the read concern header if it isn't set to balanced.

Make the base cache test class create a client with either balanced or
consistent read concern based on whether the CONSISTENT_READS
env var is defined.

Add tasks to run integration tests for the different clients.

Add make targets to use the new tasks.

Organize integration tests into auth/cache/storage/topics packages.