chore: consolidate gRPC stubs handling logic into an abstract class
Create an abstract gRPC stubs manager that the existing stubs managers
can descend from. Almost all the logic in the stubs managers is shared,
so it should go into one place. Retries and the read concern header are
moved into the common logic, even though they are only used in the cache
data manager, to make them easier to use by other managers in the
future.

Make the close method in the stubs managers await termination of the
gRPC channels and retry executors. This ensures they will try to
complete any in-flight calls before shutting down.

Propagate the thread interruption flag in the base internal client close
method to make it work better with other close methods.

Add withDeadline to the grpc configuration interface so it can be used
in the stubs managers.