Add docs and address comments.
Add docs to ExponentialBackoffRetryStrategy.

Add new constructors to ExponentialBackoffRetryStrategy: one that uses
defaults for initial delay and max backoff, and one that uses the
default eligibility strategy.

Make computePreviousBaseDelay return initial delay if currentBaseDelay
is equal to the initial delay to prevent it from returning 0.

Use the default eligibility strategy in the exponential backoff
integration tests.

Add a new integration test to show that ExponentialBackoffRetryStrategy
will work if there is a temporary network outage.

Rename the testNonRetryEligible temporary outage fixed delay test to
testRetryEligible, because it is testing an eligible method.