chore: improve subscription reconnection logic (#435)
Move the topic subscription reconnection logic into a retry strategy
and eligibility strategy that can be configured by the client.

Add the subscription retry strategy to TopicConfiguration. Add new
constructors that take the retry strategy, and new ones that don't take
a logger, which is only used inside the TopicConfiguration itself.

Prevent potential race conditions that could occur when unsubscribing
from a connection while it is reconnecting by checking if the user
called unsubscribe before reconnecting, and by making the variables
used in reconnection atomic.

Call close in SubscriptionWrapper when a stream ends to stop a
subscription from leaking threads from the retry executor service. We
may want to share a thread pool between the subscriptions instead of
making a new one per subscription, so that we don't need to make a new
one each time. We could also use it to execute the callbacks, which are
currently executed in the gRPC thread pool and could cause backpressure
if they are long-running.

Add empty default implementations for the test methods in
IScsTopicConnection to clean up its use in the topic client.

Remove SendSubscribeOptions, since we already have a class that contains
the callbacks, and because it had an unused subscription object.

Get rid of CancelableClientCallStreamObserver's ClientCallStreamObserver
extension, because that class is meant for observing requests, not
responses, and we only need a simple cancel method.

Make TopicClientLocalTest's unrecoverable error 'not found', because the
previous error is now recoverable.

Unsubscribe should end with onCompleted instead of onError

---------

Co-authored-by: anitarua <anita@momentohq.com>