fix: add bookkeeping for number of active subscriptions (#448)
Found that the java sdk does not keep track of the number of concurrent subscriptions while auditing the implementations of separate grpc channels for publish and subscribe.
Also added a new retry test file to verify that the TopicClient does not allow >100 streams on a single grpc channel and that it is capable of handling bursts of subscribe and unsubscribe requests.